Move train_step into pix2pix
